Change Vehicle dimensions to float and add overwriteReservation
Updated the Vehicle struct in reservation.h to use float for height and length instead of double. Added a new static method overwriteReservation to the ReservationASM class in reservationASM.h for overwriting reservations at a specific index.

+++ reservation.h
@@ -7,8 +7,8 @@
  * 
 */
 struct Vehicle {
-    double height;
-    double length;
+    float height;
+    float length;
 };
 
 class Reservation {
